installation script fixes

+++ install.sh
@@ -15,16 +15,14 @@ files="bash_profile bashrc vimrc vim dircolors gitconfig"    # list of files/fol
 # create dotfiles_old in homedir
 echo "Creating $olddir for backup of any existing dotfiles in ~"
 mkdir -p $olddir
-echo "...done"
 
 # change to the dotfiles directory
 echo "Changing to the $dir directory"
 cd $dir
-echo "...done"
 
+echo "Moving any existing dotfiles from ~ to $olddir, and placing new ones in their place..."
 for file in $files; do
-  echo "Moving any existing dotfiles from ~ to $olddir"
-  mv ~/.$file ~/dotfiles_old/
-  echo "Creating symlink to $file in home directory."
+  echo "  $file"
+  [[ -e ~/.$file ]] && mv ~/.$file ~/dotfiles_old/
   ln -s $dir/$file ~/.$file
 done
